code cleanup, made debug stuff a #define option
